test(node): move `https` tests
